Manage nil values in isPrimitiveSetMap

+++ src/de.hpi.swa.graal.squeak/src/de/hpi/swa/graal/squeak/nodes/accessing/SqueakObjectAt0Node.java
@@ -38,7 +38,7 @@ public abstract class SqueakObjectAt0Node extends AbstractNode {
     @Specialization
     protected static final Object doArray(final ArrayObject obj, final long index,
                     @Cached final ArrayObjectReadNode readNode) {
-        return readNode.execute(obj, index);
+        return readNode.execute(obj, (int) index);
     }
 
     @Specialization
